added copytotclipboard component
