+++ src/holofuel/components/CopyToClipboard/CopyToClipboard.module.css
@@ -1,0 +1,3 @@
+.copy-item {
+  cursor: copy;
+}
