+++ src/holofuel/components/CopyToClipboard/index.js
@@ -1,0 +1,3 @@
+import component from './CopyToClipboard'
+
+export default component
