+++ src/holofuel/components/HashIcon/HashIcon.js
@@ -1,7 +1,5 @@
 import { string, number } from 'prop-types'
-import React, { useState } from 'react'
-import ReactTooltip from 'react-tooltip'
-import copy from 'copy-to-clipboard'
+import React from 'react'
 import { Network as Identicon } from 'react-identicon-variety-pack'
 import './HashIcon.module.css'
 
@@ -10,23 +8,11 @@ export default function HashIcon ({
   size = 64,
   className
 }) {
-  const [copied, setCopied] = useState(false)
-  const copyHash = () => {
-    copy(hash)
-    setCopied(true)
-  }
-
-  return <div data-tip={hash} onClick={copyHash} data-testid='hash-icon'>
-    <Identicon
-      seed={hash}
-      size={size}
-      className={className}
-      circle />
-    <ReactTooltip
-      delayShow={250}
-      afterHide={() => setCopied(false)}
-      getContent={dataTip => dataTip === hash ? `${hash} - ${copied ? 'Copied to clipboard' : 'Click to copy'}` : null} />
-  </div>
+  return <Identicon
+    seed={hash}
+    size={size}
+    className={className}
+    circle />
 }
 
 HashIcon.propTypes = {
