+++ src/holofuel/pages/CreateOffer/CreateOffer.js
@@ -7,8 +7,8 @@ import Loader from 'react-loader-spinner'
 import HolofuelOfferMutation from 'graphql/HolofuelOfferMutation.gql'
 import HolofuelCounterpartyQuery from 'graphql/HolofuelCounterpartyQuery.gql'
 import PrimaryLayout from 'holofuel/components/layout/PrimaryLayout'
-import useFlashMessageContext from 'holofuel/contexts/useFlashMessageContext'
 import HashIcon from 'holofuel/components/HashIcon'
+// import useFlashMessageContext from 'holofuel/contexts/useFlashMessageContext'
 import Button from 'holofuel/components/Button'
 import './CreateOffer.module.css'
